refactor(db): Move MITM protocol submission into dedicated class
* Add PogoProtoSubmit class
* Simplify method names
* Speed up quest_proto processing logic
* Simplify spawndef logic
